pyproject.toml
[project]
name = "okx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "python-dotenv>=1.1.0",
    "python-okx>=0.3.9",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
